chore: Remove unnecessary rollup plugins
Because I want to treat all NPM packages as external, it is not
necessary for rollup to bundle or find any NPM packages.

Because I do not want rollup to bundle, and therefore have to parse
any NPM packages, I do not need the CommonJS rollup plugin.

Because I do not need rollup to find any NPM packages (in order to
bundle them) I do not need to teach it the node resolve alrogithm, and
thereefore don't need the node reosolve plugin.